refactor: SSA 구조 적용(SignIn, SignUp) closes #3
- startDestination 동적 설정
- 직렬화된 route 대신 string 기반 route 비교 적용
- 회원가입 후 로그인 화면에서 Toast 메시지 표시 로직 추가
