--- app/src/main/java/org/sopt/at/presentation/main/screen/ShortsScreen.kt
@@ -1,46 +1,0 @@
-package org.sopt.at.presentation.main.screen
-
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.sp
-import org.sopt.at.R
-import org.sopt.at.ui.theme.TvingTheme
-
-@Composable
-fun ShortsScreen(
-    paddingValues: PaddingValues,
-    modifier: Modifier = Modifier,
-) {
-    Column(
-        modifier = modifier
-            .fillMaxSize()
-            .background(Color.Black),
-        verticalArrangement = Arrangement.Center,
-        horizontalAlignment = Alignment.CenterHorizontally,
-    ) {
-        Text(
-            text = stringResource(R.string.shorts_title),
-            color = Color.White,
-            fontSize = 24.sp,
-        )
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-private fun Preview() {
-    TvingTheme {
-        ShortsScreen(paddingValues = PaddingValues())
-    }
-
-}
