+++ app/src/main/res/values/strings.xml
@@ -84,7 +84,7 @@
     <string name="btn_basic_topbar_back">뒤로가기</string>
     <string name="btn_password_visibility">password visibility button</string>
     <string name="btn_logout">로그아웃</string>
-    <string name="success_signin_message">회원가입 성공!</string>
+    <string name="success_signup_message">회원가입 성공!</string>
     <string name="title_activity_main">MainActivity</string>
     <string name="tving_logo">TVING Logo</string>
     <string name="my_profile_button">My profile button</string>
